Commit no.4: Implementing almost all of the remaining methods. 3 tests pass.

+++ helpers/Modificator.java
@@ -2,5 +2,5 @@ package helpers;
 
 public interface Modificator {
     float accept(ModificatorVisitor modificatorVisitor, float[] modificators);
-    float accept(ModificatorVisitor modificatorVisitor, int[] hp);
+    int accept(ModificatorVisitor modificatorVisitor, int[] hp);
 }
